Test program to measure process creation/destruction
Process destruction on Windows 10 seems to be serialized and to get
worse over time. This tests this by creating processes as quickly as
possible, waiting 0.5 s, and then destroying the processes as quickly
as possible. In addition to measuring the time required for creation
and destruction, access to the lock which is required to call
PeekMessage and many other functions is tested.