Add uptime display to ProcessCreateTests
Process destruction lock contention on Windows 10 seems to get worse
over time, presumably as the system is used more. This can lead to false
reports of fixes because installing a new version of Windows 10 always
involves a reboot, and any reboot improves process destruction.

To help manage this problem this change adds printing of uptime in days.
The uptime is printed in two forms - time since the last reboot, and
non-sleep time since the last reboot. The former is what task manager
shows but the latter is presumably more relevant.

This also includes a new ProcessCreateTests.exe.