Add tool for finding zombie processes
If one process gets a handle to another process - through CreateProcess
or OpenProcess or other methods - and fails to close than handle then
when the other process dies it becomes a zombie. The kernel cannot
cleanup all of its data until the last handle is closed. Leaking a few
handles is harmless, but some programs leak thousands.

This tool uses NtApiDotNet to find zombie processes, and who owns their
handles. If you see a program that you use creating zombies, file a bug.
If you see a program that you wrote creating zombies, fix it.