Find CFG reservation, use SetProcessValidCallTargets
In order to experiment with SetProcessValidCallTargets and see what bits
it sets in the CFG reservation I wrote a trivial loop that finds the CFG
reservation, used SetProcessValidCallTargets to mark some branch targets
in the freshly allocated memory as being valid, and then used the
debugger to understand what is going on.

I now allocate the memory with PAGE_TARGETS_INVALID to make it clear
that all branch targets are invalid until marked as valid with
SetProcessValidCallTargets. Only 16-byte aligned offsets are allowed so
only odd-numbered bits will be set.

If you halt after the call to SetProcessValidCallTargets the first byte
of this expression will be 0x55:

    (size_t)p/64 + (unsigned char*)cfg_reservation,1000x

Optimizations are disabled to make it easier to use the debugger to view
CFG memory - otherwise the cfg_reservation variable is optimized away.

Calling SetProcessValidCallTargets was challenging. The import library
was not obvious, but I found it (mincore.lib). If you want to
dynamically call it - for code that has to work on earlier OSes, you
have to call it through a function pointer and that hits a (wait for
it...) CFG violation. So I had to tag the wrapper function with
DECLSPEC_GUARDNOCF.