Print CFG self-stats for the VAllocStress process
A process can easily scan it's own virtual address space to find its own
CFG block and count how many blocks are there. This modifies
VirtualStress so that it finds its CFG reservation and then prints how
many blocks are initially allocated. It then does its stressful
allocations and then tries to print the new CFG block count. Because CFG
scanning may be very slow it may not finish counting. It uses the CFG
address from the initial scan (which will always be fast) to allow this
estimate to work. The committed byte count is also printed, for easy
sanity checking of the allocation claims.