Allow optionally loading user32.dll into all processes
ProcessCreateTests demonstrates how user-crit lock contention can make
process destruction on Windows very slow. With this change it is easy to
demonstrate how process destruction on Windows is *even* slower if
user32.dll or gdi32.dll are loaded. Loading these DLLs triggers the
creation of GDI objects and these are destroyed at process destruction
time and this requires acquiring the user-crit and this can easily cause
the mouse to lock up for a second or two, even on a new laptop running
the latest version of Windows 10 with less than a day of awake uptime.

Everything is horrible.