Use SetProcessMitigationPolicy to avoid GDI costs
This change adds (commented out) code that uses
SetProcessMitigationPolicy to avoid initializing GDI, thereby avoiding
the costs of both initializing and tearing down GDI, and the associated
hangs. This must be done before GDI is initialized, which typically
means it needs to be done together with /delayload, but it gives what
appears to be an effective alternative if some user32.dll/shell32.dll
functions are genuinely needed.

This suggestion came from here:
https://old.reddit.com/r/programming/comments/a2y16c/a_notcalled_function_can_cause_a_5x_slowdown/eb29x6z/