Use-after-unload sample for !analyze testing
!analyze failed to find a use-after-unload bug so I wanted to test and
see if this was specific to that crash, or a systemic problem. It looks
like !analyze can detect use-after-unload crashes if the instruction
pointer is in the unloaded DLL, but if it is a data reference in an
unloaded DLL then all bets are off. That is, when making a function
call to an unloaded DLL !analyze says this:

Probably caused by : class.dll ( class.dll!Unloaded+2a81 )

However when incrementing a variable in an unloaded DLL !analyze says
this:

Probably caused by : main.exe ( main!main+12a )

Similarly, making a virtual function call to an unloaded DLL will tend
to fail because the first step is loading an address from a v-table in
the DLL.

TL;DR - !analyze can detect use-after-unload problems only when the
instruction pointer is in the unloaded DLL.