Add multitest.bat to repeatedly create processes
CreateProcessTiming.exe measures how long it takes to call CreateProcess
but to be scientific we need to do this repeatedly. This adds a batch
file that calls CreateProcessTiming.exe on itself twenty times.

With Windows Security/Defender's Real-time protection enabled this takes
2.9 to 3.4 ms per call. With it disabled this takes 1.4 to 1.8 ms per
call, roughly. So, about 1.5 ms of overhead from Windows Security.

This is, to be clear, with CreateProcessTiming.exe in an excluded
directory, so the overhead should, one would think, be zero.