Add tests for windbg handling of noreturn
As documented in this blog post:
https://randomascii.wordpress.com/2017/12/10/analyzing-a-confusing-crash/
certain code-gen patterns expose odd handling of return addresses in
windbg. This problem used to also exists in Visual Studio but was fixed.

The problem is that when walking the call stack windbg subtracts one
from addresses to ensure that its symbol lookups are within the
function. This is fine, but windbg sometimes then uses this decremented
address to do disassembly, and this confuses it.

This can be seen by loading noreturn.dmp into windbg, or by running
noreturn.exe in windbg, or by running build.bat to do the end-to-end
process.