Delayload failure demonstration
This simple project demonstrates an executable that links to a DLL and
uses /delayload so that the executable can run when the DLL is not in
the path. Then when the executable runs the DLL is not found, leading to
a cryptic failure.

This project was created to demonstrate the most methodical way of
investigating these problems.