Add test_new_branch.bat
Add a batch file for easy comparing of build times for Chrome. It does a
full build of Chrome (defaulting to 24-way parallelism, tweak as needed)
first from origin/master, and then from the specified branch. An
optional revision number argument can be specified to give more context
for the results, but the revision number specified does not affect what
code is built.