Improve counting and measuring of interrupts
count_interrupts.cpp counts how many timer interrupts have happened by
spinning in a busy loop (!!!) calling timeGetTime() and recording
whenever its return value changes. It also prints out a histogram of
the deltas.

measure_interrupts.cpp was updated so that it prints a histogram of
Sleep(1) delays (with commented out code to print all of the delays
when the timer interrupt interval is set to 8).