Trigger FindZombieHandles failure
When ProcessCreateTests is run with these changes and FindZombieHandles
is run (after ProcessCreateTests prints "Waiting...") then
FindZombieHandles fails when an exception is thrown. It prints:

(0xC0000023) - {Buffer Too Small}
The buffer is too small to contain the entry. No information has been written to the buffer.

It fails in GetZombieProcessObjectAddress on the NtSystemInfo.GetHandles line.