Improvements to PopulateDesktop.py
This improves the PopulateDesktop.py script to make it easier to blog
about and to get cleaner data. The biggest change is to use psutil
to suspend explorer.exe while populating the desktop with new images.

This also includes a batch file to simplify running the tests. The
tester is only required to monitor explorer.exe CPU usage so that they
know when to move to the next test.