Show that LoadLibrary leaks CFG memory
It was found years ago that allocating and then freeing executable
memory in a CFG enabled process caused the associated CFG commit to be
leaked. This change adds testing of CFG commit leaking by repeatedly
loading and unloading a DLL leaks CFG entries. The answer is yes,
although to really see it you need to force the DLL to a different
address every time.

This change does that by spinning in a loop loading and then freeing
msvcrt.dll and then allocating some memory where it used to be. Details
are here:

https://randomascii.wordpress.com/2018/08/16/24-core-cpu-and-i-cant-type-an-email-part-one/