Add zeropage_test.cpp
This source file should be compiled with "cl zeropage_test.cpp" to
create an executable called zeropage_test.exe. When this process is run
it will run another copy of itself in order to increase the allocation
load on the system. On Windows 10-based operating systems (including
Windows Server 2022) on machines with 96+ cores this triggers
pathological behavior in the System process where the
MiZeroLargePageThread threads spend most of their time fighting over a
spin lock instead of zeroing memory.