Test code for PROCESS_MODE_BACKGROUND_BEGIN
When a process is put into background mode with this command:
  SetPriorityClass(GetCurrentProcess(), PROCESS_MODE_BACKGROUND_BEGIN);

then the following happens, per Microsoft's documentation:

  Begin background processing mode. The system lowers the
  resource scheduling priorities of the process (and its threads) so
  that it can perform background work without significantly affecting
  activity in the foreground.
  This value can be specified only if hProcess is a handle to the
  current process. The function fails if the process is already in
  background processing mode.

It's a big vague but the overall idea is clear - this should lower the
impact of the process on the system.

However, this can be hugely inaccurate. One effect of this call is to
limit the working set of the process to 32 MiB. If the process touches
more memory than this then the kernel will repeatedly fault in pages and
trim the working set, causing work to take about 300 times as long to
run. In other words, the CPU load in this scenario will be dramatically
_increased_.

Therefore, this call should never be used.

This test code demonstrates this problem. Run it with and without the
"background" parameter to see the difference in performance.