Improve background test code
This change improves the background test code so that it runs repeatedly
in normal mode and then background mode. It always runs for about 1,000
ms (rather than needing configuration variables) and after each
background-mode run it prints the ratio. Thus, compiling and running
this provides an estimate of the slowdown caused by background mode in
this test case. It is... considerable. And surprisingly variable.