Framework update and warn about handle leaks from drivers
Drivers can leak handles and this can lead to process zombies whose
cause is not detectable by this tool. This change updates this tool to
point out this possibility. If the number of zombie process is far
higher than can be explained by handle leaks in processes it suggests
that a driver is the culprit and prints a link to an explanatory blog
post.

I also now print a link to the original blog post.