Properly publish FindZombieHandles to the prebuilt folder
This publishes FindZombieHandles using the Build-> Publish command as
explained here:

https://learn.microsoft.com/en-us/dotnet/core/tutorials/publishing-with-visual-studio

It also fixes a bug in the previous publish-commit - the .exe and .pdb was not
added because *.exe is in the root-level .gitignore file.

Running the executable requires installing the .Net 9.0 Runtime using
this link:

https://dotnet.microsoft.com/en-us/download/dotnet/thank-you/runtime-9.0.11-windows-x64-installer?cid=getdotnetcore

The publishing settings used are now recorded in FolderProfile.pubxml
and the local .gitignore file has been updated to prevent future
publishing errors.